Factor shared methods of Receivers and Methods out
