+++ kmock-processor/src/main/kotlin/tech/antibytes/kmock/processor/KMockProcessorProvider.kt
@@ -33,6 +33,7 @@ import tech.antibytes.kmock.processor.mock.KMockProxyNameSelector
 import tech.antibytes.kmock.processor.mock.KMockReceiverGenerator
 import tech.antibytes.kmock.processor.mock.KMockRelaxerGenerator
 import tech.antibytes.kmock.processor.mock.KMockSpyGenerator
+import tech.antibytes.kmock.processor.mock.MethodeGeneratorHelper
 import tech.antibytes.kmock.processor.multi.KMockMultiInterfaceBinder
 import tech.antibytes.kmock.processor.multi.KMockParentFinder
 import tech.antibytes.kmock.processor.utils.AnnotationFilter
@@ -145,13 +146,17 @@ class KMockProcessorProvider(
             nameSelector = nameSelector,
             nonIntrusiveInvocationGenerator = nonIntrusiveInvocationGenerator,
         )
+        val methodeGeneratorHelper = MethodeGeneratorHelper(
+            genericResolver = KMockGenerics,
+        )
         val methodGenerator = KMockMethodGenerator(
+            utils = methodeGeneratorHelper,
             nameSelector = nameSelector,
             nonIntrusiveInvocationGenerator = nonIntrusiveInvocationGenerator,
             genericResolver = KMockGenerics,
         )
-
         val receiverGenerator = KMockReceiverGenerator(
+            utils = methodeGeneratorHelper,
             nameSelector = nameSelector,
             genericResolver = KMockGenerics,
         )
