+++ kmock-processor/src/main/kotlin/tech/antibytes/kmock/processor/mock/KMockMethodGenerator.kt
@@ -7,30 +7,18 @@
 package tech.antibytes.kmock.processor.mock
 
 import com.google.devtools.ksp.symbol.KSFunctionDeclaration
-import com.google.devtools.ksp.symbol.KSType
-import com.google.devtools.ksp.symbol.KSTypeParameter
 import com.google.devtools.ksp.symbol.KSTypeReference
-import com.google.devtools.ksp.symbol.KSValueParameter
 import com.google.devtools.ksp.symbol.Modifier
 import com.squareup.kotlinpoet.AnnotationSpec
-import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
-import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeName
-import com.squareup.kotlinpoet.TypeVariableName
-import com.squareup.kotlinpoet.asClassName
-import com.squareup.kotlinpoet.asTypeName
 import com.squareup.kotlinpoet.ksp.TypeParameterResolver
 import com.squareup.kotlinpoet.ksp.toTypeName
 import com.squareup.kotlinpoet.ksp.toTypeParameterResolver
-import com.squareup.kotlinpoet.ksp.toTypeVariableName
-import tech.antibytes.kmock.KMockContract.AsyncFunProxy
-import tech.antibytes.kmock.KMockContract.SyncFunProxy
-import tech.antibytes.kmock.processor.ProcessorContract.GenericDeclaration
+import tech.antibytes.kmock.processor.ProcessorContract.MethodeGeneratorHelper
 import tech.antibytes.kmock.processor.ProcessorContract.GenericResolver
-import tech.antibytes.kmock.processor.ProcessorContract.MethodArgumentTypeInfo
 import tech.antibytes.kmock.processor.ProcessorContract.MethodGenerator
 import tech.antibytes.kmock.processor.ProcessorContract.MethodReturnTypeInfo
 import tech.antibytes.kmock.processor.ProcessorContract.MethodTypeInfo
@@ -38,269 +26,13 @@ import tech.antibytes.kmock.processor.ProcessorContract.NonIntrusiveInvocationGe
 import tech.antibytes.kmock.processor.ProcessorContract.ProxyInfo
 import tech.antibytes.kmock.processor.ProcessorContract.ProxyNameSelector
 import tech.antibytes.kmock.processor.ProcessorContract.Relaxer
-import tech.antibytes.kmock.processor.utils.toSecuredTypeName
 
 internal class KMockMethodGenerator(
+    private val utils: MethodeGeneratorHelper,
     private val nameSelector: ProxyNameSelector,
     private val nonIntrusiveInvocationGenerator: NonIntrusiveInvocationGenerator,
     private val genericResolver: GenericResolver,
 ) : MethodGenerator {
-    private data class ProxyBundle(
-        val proxy: PropertySpec,
-        val returnType: MethodReturnTypeInfo
-    )
-
-    private fun determineArguments(
-        inherited: Boolean,
-        arguments: List<KSValueParameter>,
-        typeParameterResolver: TypeParameterResolver
-    ): Array<MethodTypeInfo> {
-        return arguments.map { parameter ->
-            val argumentName = parameter.name!!.asString()
-
-            parameter.type.modifiers
-            MethodTypeInfo(
-                argumentName = argumentName,
-                typeName = parameter.type.toSecuredTypeName(
-                    inheritedVarargArg = parameter.isVararg && inherited,
-                    typeParameterResolver = typeParameterResolver
-                ),
-                isVarArg = parameter.isVararg,
-            )
-        }.toTypedArray()
-    }
-
-    private fun determineTypeParameter(
-        parameter: List<KSTypeParameter>,
-        typeParameterResolver: TypeParameterResolver
-    ): List<TypeName> {
-        var distribute = false
-        val parameterTypes = parameter.map { type ->
-            val parameterType = type.toTypeVariableName(typeParameterResolver)
-
-            if (parameterType.bounds.size > 1) {
-                distribute = true
-            }
-
-            parameterType
-        }
-
-        return if (distribute) {
-            parameterTypes
-        } else {
-            emptyList()
-        }
-    }
-
-    private fun determineProxyType(suspending: Boolean): Triple<ClassName, String, String> {
-        return if (suspending) {
-            Triple(asyncProxy, "createAsyncFunProxy", "suspend ")
-        } else {
-            Triple(syncProxy, "createSyncFunProxy", "")
-        }
-    }
-
-    private fun resolveProxyGenerics(
-        generics: Map<String, List<KSTypeReference>>?,
-        typeResolver: TypeParameterResolver
-    ): Map<String, GenericDeclaration>? {
-        return if (generics == null) {
-            null
-        } else {
-            genericResolver.mapProxyGenerics(generics, typeResolver)
-        }
-    }
-
-    private fun resolveGenericProxyType(
-        typeName: TypeName,
-        generic: GenericDeclaration?
-    ): TypeName {
-        return when {
-            generic == null -> typeName
-            generic.types.size > 1 -> any.copy(nullable = generic.nullable)
-            else -> generic.types.first()
-        }
-    }
-
-    private fun mapGenericProxyType(
-        typeName: TypeName,
-        classScopeGenerics: Map<String, List<TypeName>>?,
-        proxyGenericTypes: Map<String, GenericDeclaration>,
-    ): MethodReturnTypeInfo {
-        val isNullable = typeName.isNullable
-        val generic = proxyGenericTypes[typeName.toString().trimEnd('?')]
-
-        val actualTypeName = resolveGenericProxyType(
-            typeName = typeName,
-            generic = generic
-        )
-
-        return MethodReturnTypeInfo(
-            typeName = typeName,
-            actualTypeName = actualTypeName.copy(nullable = actualTypeName.isNullable || isNullable),
-            generic = generic,
-            classScope = classScopeGenerics
-        )
-    }
-
-    private fun determineProxyReturnType(
-        returnType: KSType,
-        classScopeGenerics: Map<String, List<TypeName>>?,
-        proxyGenericTypes: Map<String, GenericDeclaration>?,
-        typeResolver: TypeParameterResolver,
-    ): MethodReturnTypeInfo {
-        val typeName = returnType.toTypeName(typeResolver)
-
-        return if (proxyGenericTypes == null) {
-            MethodReturnTypeInfo(
-                typeName = typeName,
-                actualTypeName = typeName,
-                generic = null,
-                classScope = classScopeGenerics
-            )
-        } else {
-            mapGenericProxyType(
-                typeName = typeName,
-                classScopeGenerics = classScopeGenerics,
-                proxyGenericTypes = proxyGenericTypes,
-            )
-        }
-    }
-
-    private fun resolveGenericArgumentTypes(
-        proxyGenericTypes: Map<String, GenericDeclaration>,
-        classScopeGenerics: Map<String, List<TypeName>>?,
-        argumentTypes: Array<MethodTypeInfo>,
-    ): List<MethodArgumentTypeInfo> {
-        return argumentTypes.map { typeInfo ->
-            val (_, actualTypeName, declaration, _) = mapGenericProxyType(
-                typeName = typeInfo.typeName,
-                classScopeGenerics = classScopeGenerics,
-                proxyGenericTypes = proxyGenericTypes,
-            )
-
-            MethodArgumentTypeInfo(
-                typeInfo = typeInfo.copy(typeName = actualTypeName),
-                generic = declaration
-            )
-        }
-    }
-
-    private fun determineProxyArgumentTypes(
-        proxyGenericTypes: Map<String, GenericDeclaration>?,
-        classScopeGenerics: Map<String, List<TypeName>>?,
-        argumentTypes: Array<MethodTypeInfo>,
-    ): List<MethodArgumentTypeInfo> {
-        return if (proxyGenericTypes == null) {
-            argumentTypes.map { typeInfo ->
-                MethodArgumentTypeInfo(
-                    typeInfo = typeInfo,
-                    generic = null,
-                )
-            }
-        } else {
-            resolveGenericArgumentTypes(
-                proxyGenericTypes = proxyGenericTypes,
-                classScopeGenerics = classScopeGenerics,
-                argumentTypes = argumentTypes,
-            )
-        }
-    }
-
-    private fun mapProxyArgumentTypeNames(
-        proxyArguments: List<MethodArgumentTypeInfo>
-    ): List<TypeName> {
-        return proxyArguments.map { argument ->
-            if (!argument.typeInfo.isVarArg) {
-                argument.typeInfo.typeName
-            } else {
-                specialArrays.getOrElse(argument.typeInfo.typeName) {
-                    TypeVariableName(
-                        "Array<out ${argument.typeInfo.typeName}>"
-                    )
-                }
-            }
-        }
-    }
-
-    private fun buildSideEffectSignature(
-        proxyArguments: List<MethodArgumentTypeInfo>,
-        proxyReturnType: TypeName,
-        prefix: String
-    ): TypeName {
-        val argumentTypeName = mapProxyArgumentTypeNames(proxyArguments)
-
-        return TypeVariableName(
-            "$prefix(${argumentTypeName.joinToString(", ")}) -> $proxyReturnType"
-        )
-    }
-
-    private fun buildProxyInitializer(
-        proxySpec: PropertySpec.Builder,
-        proxyInfo: ProxyInfo,
-        proxyFactoryMethod: String,
-    ): PropertySpec.Builder {
-        return proxySpec.initializer(
-            "ProxyFactory.%L(%S, collector = verifier, freeze = freeze)",
-            proxyFactoryMethod,
-            proxyInfo.proxyId,
-        )
-    }
-
-    private fun buildProxy(
-        methodScope: TypeName?,
-        proxyInfo: ProxyInfo,
-        arguments: Array<MethodTypeInfo>,
-        suspending: Boolean,
-        classScopeGenerics: Map<String, List<TypeName>>?,
-        generics: Map<String, List<KSTypeReference>>?,
-        returnType: KSType,
-        typeResolver: TypeParameterResolver,
-    ): ProxyBundle? {
-        return if (methodScope == null) {
-            val (proxyType, proxyFactoryMethod, sideEffectPrefix) = determineProxyType(suspending)
-            val proxyGenericTypes = resolveProxyGenerics(
-                generics = generics,
-                typeResolver = typeResolver
-            )
-
-            val proxyArguments = determineProxyArgumentTypes(
-                proxyGenericTypes = proxyGenericTypes,
-                classScopeGenerics = classScopeGenerics,
-                argumentTypes = arguments,
-            )
-
-            val proxyReturnType = determineProxyReturnType(
-                returnType = returnType,
-                classScopeGenerics = classScopeGenerics,
-                proxyGenericTypes = proxyGenericTypes,
-                typeResolver = typeResolver
-            )
-
-            val sideEffect = buildSideEffectSignature(
-                proxyArguments = proxyArguments,
-                proxyReturnType = proxyReturnType.actualTypeName,
-                prefix = sideEffectPrefix
-            )
-
-            ProxyBundle(
-                PropertySpec.builder(
-                    proxyInfo.proxyName,
-                    proxyType.parameterizedBy(proxyReturnType.actualTypeName, sideEffect)
-                ).let { proxySpec ->
-                    buildProxyInitializer(
-                        proxySpec = proxySpec,
-                        proxyInfo = proxyInfo,
-                        proxyFactoryMethod = proxyFactoryMethod,
-                    )
-                }.build(),
-                proxyReturnType
-            )
-        } else {
-            null
-        }
-    }
-
     private fun FunSpec.Builder.addArguments(
         arguments: Array<MethodTypeInfo>
     ): FunSpec.Builder {
@@ -420,12 +152,12 @@ internal class KMockMethodGenerator(
         val typeParameterResolver = ksFunction.typeParameters
             .toTypeParameterResolver(typeResolver)
         val generics = genericResolver.extractGenerics(ksFunction, typeParameterResolver)
-        val arguments = determineArguments(
+        val arguments = utils.determineArguments(
             inherited = inherited,
             arguments = ksFunction.parameters,
             typeParameterResolver = typeParameterResolver
         )
-        val parameter = determineTypeParameter(
+        val parameter = utils.determineTypeParameter(
             parameter = ksFunction.typeParameters,
             typeParameterResolver = typeParameterResolver,
         )
@@ -440,8 +172,7 @@ internal class KMockMethodGenerator(
         val returnType = ksFunction.returnType!!.resolve()
         val isSuspending = ksFunction.modifiers.contains(Modifier.SUSPEND)
 
-        val proxySignature = buildProxy(
-            methodScope = methodScope,
+        val proxySignature = utils.buildProxy(
             proxyInfo = proxyInfo,
             arguments = arguments,
             classScopeGenerics = classScopeGenerics,
@@ -460,37 +191,18 @@ internal class KMockMethodGenerator(
             enableSpy = enableSpy,
             parameter = parameter,
             arguments = arguments,
-            proxyReturnType = proxySignature?.returnType,
+            proxyReturnType = proxySignature.returnType,
             typeResolver = typeParameterResolver,
             relaxer = relaxer
         )
 
-        return Pair(proxySignature?.proxy, method)
+        return Pair(proxySignature.proxy, method)
     }
 
     private companion object {
-        private val any = Any::class.asTypeName()
         private val unchecked = AnnotationSpec.builder(Suppress::class).addMember("%S", "UNCHECKED_CAST").build()
         private val varargs = arrayOf(KModifier.VARARG)
         private val noVarargs = arrayOf<KModifier>()
-        private val asyncProxy = AsyncFunProxy::class.asClassName()
-        private val syncProxy = SyncFunProxy::class.asClassName()
         private val scopedBody = "throw IllegalStateException(\n\"This action is not callable.\"\n)"
-
-        @OptIn(ExperimentalUnsignedTypes::class)
-        private val specialArrays: Map<TypeName, TypeName> = mapOf(
-            Int::class.asTypeName() to IntArray::class.asTypeName(),
-            Byte::class.asTypeName() to ByteArray::class.asTypeName(),
-            Short::class.asTypeName() to ShortArray::class.asTypeName(),
-            Long::class.asTypeName() to LongArray::class.asTypeName(),
-            Float::class.asTypeName() to FloatArray::class.asTypeName(),
-            Double::class.asTypeName() to DoubleArray::class.asTypeName(),
-            Char::class.asTypeName() to CharArray::class.asTypeName(),
-            Boolean::class.asTypeName() to BooleanArray::class.asTypeName(),
-            UByte::class.asTypeName() to UByteArray::class.asTypeName(),
-            UShort::class.asTypeName() to UShortArray::class.asTypeName(),
-            UInt::class.asTypeName() to UIntArray::class.asTypeName(),
-            ULong::class.asTypeName() to ULongArray::class.asTypeName(),
-        )
     }
 }
